+++ xdeps/madxutils.py
@@ -121,6 +121,11 @@ class MadxEnv:
              'elements':self._elements,
              'xdeps':self.manager.dump()}
 
+    def reload(self,data):
+        self._variables.update(data['variables'])
+        self._elements.update(data['elements'])
+        self.manager.reload(data['xdeps'])
+
     def to_json(self,filename):
         import json
         json.dump(self.dump(),open(filename,'w'))
@@ -129,10 +134,7 @@ class MadxEnv:
     def from_json(cls,filename):
         import json
         self=cls()
-        data=json.load(open(filename))
-        self._variables.update(data['variables'])
-        self._elements.update(data['elements'])
-        self.manager.reload(data['xdeps'])
+        self.reload(json.load(open(filename)))
         return self
 
     def read_state(self,mad):
